Update regex to also substitute the longest line
Previously we excluded lines that had the same length to the first `:`
as the longest line, this left us with comments that were off by spaces
since we weren't re-writing those as well. While fixing this bug I added
an assert that the lines matched, they should always match and re.sub
returns the same string if there are no matches.